An web application built using XML
It satisfies the following requirements.
1.All xml instances must validate against their dtd.
2.Due in the Recipe Project drop boxbefore you start presenting the day
of the presentation.
3.Must have 3 separate XML files one for each recipe to demonstrate the
process.
4.Each Recipe XML file must include a reference to a general entity
thatwill be an XMLfilethat is a portion of the content.An example you
might split your recipe up into ingredients' and instructions' files for
convenience.
5.No namespace will be needed for this project.
6.Do not use attributes for textual information that might be displayed
when the xml file is converted to an html or pdf file.
7.Each filemust reference an image file and there will be a NOTATION
entry for it.
8.Each recipe file must have a mixed content construct usedsomewhere in
the complete instance.
9.All of the recipefiles must validateagainst the DTD. 10.Each
recipefile must use an ID/IDREF type of attribute to make a connection
between information areas. 11.There will be one DTD that will call at
lease 2 modules that are part of the DTD.
12.The DTD must use at least one parameter entity.
13.The DTD will have an enumeration (group) attribute. 14.The DTD will
have at least on empty element that will have some sensible use.
15.The DTD should be commented so maintainers will be able to do their
job after you are on another project.